+++ tools/accuracy_checker/accuracy_checker/evaluators/custom_evaluators/colorization_evaluator.py
@@ -321,6 +321,7 @@ class BaseModel:
         self.input_blob = None
         self.output_blob = None
         self.with_prefix = False
+        self.is_dynamic = False
         if not delayed_model_loading:
             self.load_model(network_info, launcher, log=True)
 
@@ -365,8 +366,7 @@ class BaseModel:
         model, weights = self.auto_model_search(network_info, self.net_type)
         if weights:
             self.network = launcher.read_network(str(model), str(weights))
-            self.network.batch_size = 1
-            self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
+            self.load_network(self.network, launcher.device)
         else:
             self.network = None
             launcher.ie_core.import_network(str(model))
@@ -376,8 +376,30 @@ class BaseModel:
 
     def load_network(self, network, launcher):
         self.network = network
-        self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
+        self.dynamic_inputs, self.partial_shapes = launcher.get_dynamic_inputs(self.network)
+        if self.dynamic_inputs and launcher.dynamic_shapes_policy in ['dynamic', 'default']:
+            try:
+                self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
+                self.is_dynamic = True
+            except RuntimeError as e:
+                if launcher.dynamic_shapes_policy == 'dynamic':
+                    raise e
+                self.is_dynamic = False
+                self.exec_network = None
+        if not self.dynamic_inputs:
+            self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
         self.set_input_and_output()
+
+    def reshape_net(self, shape):
+        if self.is_dynamic:
+            return
+        if hasattr(self, 'exec_network') and self.exec_network is not None:
+            del self.exec_network
+        self.network.reshape(shape)
+        self.dynamic_inputs, self.partial_shapes = self.launcher.get_dynamic_inputs(self.network)
+        if not self.is_dynamic and self.dynamic_inputs:
+            return
+        self.exec_network = self.launcher.load_network(self.network, self.launcher.device)
 
     def set_input_and_output(self):
         pass
@@ -404,12 +426,14 @@ class BaseModel:
         for name, input_info in network_inputs.items():
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(input_info.precision))
-            print_info('\tshape {}\n'.format(input_info.shape))
+            print_info('\tshape: {}\n'.format(
+                input_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
         print_info('{} - Output info'.format(self.net_type))
         for name, output_info in network_outputs.items():
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(output_info.precision))
-            print_info('\tshape: {}\n'.format(output_info.shape))
+            print_info('\tshape: {}\n'.format(
+                output_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
 
 
 class ColorizationTestModel(BaseModel):
@@ -445,6 +469,8 @@ class ColorizationTestModel(BaseModel):
         img_l, img_l_rs = self.data_preparation(input_data)
 
         self.inputs[self.input_blob] = img_l_rs
+        if not self.is_dynamic and self.dynamic_inputs:
+            self.reshape_net({k: v.shape for k, v in self.inputs.items()})
         res = self.exec_network.infer(inputs=self.inputs)
 
         new_result = self.postprocessing(res[self.output_blob], img_l)
@@ -456,7 +482,8 @@ class ColorizationTestModel(BaseModel):
             self.exec_network.input_info[self.input_blob].input_data
             if has_info else self.exec_network.inputs[self.input_blob]
         )
-        input_data = np.reshape(input_data, input_info.shape)
+        if not self.dynamic_inputs:
+            input_data = np.reshape(input_data, input_info.shape)
         return {self.input_blob: input_data}
 
     def set_input_and_output(self):
@@ -489,7 +516,10 @@ class ColorizationCheckModel(BaseModel):
         self.adapter.output_blob = self.output_blob
 
     def predict(self, identifiers, input_data):
-        raw_result = self.exec_network.infer(self.fit_to_input(input_data))
+        input_dict = self.fit_to_input(input_data)
+        if not self.is_dynamic and self.dynamic_inputs:
+            self.reshape_net({k: v.shape for k, v in input_dict.items()})
+        raw_result = self.exec_network.infer(input_dict)
         result = self.adapter.process([raw_result], identifiers, [{}])
         return raw_result, result
