+++ tools/accuracy_checker/accuracy_checker/evaluators/custom_evaluators/mtcnn_evaluator.py
@@ -20,11 +20,11 @@ from collections import OrderedDict
 import pickle # nosec - disable B403:import-pickle check
 from pathlib import Path
 import numpy as np
-import cv2
 
+from .mtcnn_evaluator_utils import calibrate_predictions, nms, cut_roi
 from ..base_evaluator import BaseEvaluator
 from ..quantization_model_evaluator import create_dataset_attributes
-from ...adapters import create_adapter, MTCNNPAdapter
+from ...adapters import create_adapter
 from ...launcher import create_launcher, InputFeeder
 from ...preprocessor import PreprocessingExecutor
 from ...utils import extract_image_representations, read_pickle, contains_any, get_path
@@ -308,23 +308,25 @@ class DLSDKModelMixin:
         results = []
         for feed_dict in input_blobs:
             input_shapes = {layer_name: data.shape for layer_name, data in feed_dict.items()}
-            self._reshape_input(input_shapes)
+            if not self.is_dynamic:
+                self.reshape_net(input_shapes)
             results.append(self.exec_network.infer(feed_dict))
             for meta in batch_meta:
                 meta['input_shape'].append(input_shapes)
         return results
-
-    def _reshape_input(self, input_shapes):
-        del self.exec_network
-        self.network.reshape(input_shapes)
-        self.exec_network = self.launcher.ie_core.load_network(self.network, self.launcher.device)
 
     @property
     def inputs(self):
         has_info = hasattr(self.exec_network, 'input_info')
         if not has_info:
             return self.exec_network.inputs
-        return OrderedDict([(name, data.input_data) for name, data in self.exec_network.input_info.items()])
+        inputs = OrderedDict()
+        for name, data in self.exec_network.input_info.items():
+            if name in self.partial_shapes:
+                inputs[name] = self.partial_shapes[name]
+            else:
+                inputs[name] = data.input_data
+        return inputs
 
     def input_shape(self, input_name):
         return self.inputs[input_name]
@@ -336,7 +338,9 @@ class DLSDKModelMixin:
         self.launcher.release()
 
     def fit_to_input(self, data, layer_name, layout, precision):
-        layer_shape = tuple(self.inputs[layer_name].shape)
+        layer_shape = (
+            tuple(self.inputs[layer_name].shape)
+            if layer_name not in self.dynamic_inputs else self.partial_shapes[layer_name])
         data_shape = np.shape(data)
         if len(layer_shape) == 4:
             if len(data_shape) == 5:
@@ -414,18 +418,36 @@ class DLSDKModelMixin:
 
     def load_network(self, network, launcher, model_prefix):
         self.network = network
-        self.exec_network = launcher.ie_core.load_network(network, launcher.device)
+        self.dynamic_inputs, self.partial_shapes = launcher.get_dynamic_inputs(self.network)
+        if self.dynamic_inputs and launcher.dynamic_shapes_policy in ['dynamic', 'default']:
+            try:
+                self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
+                self.is_dynamic = True
+            except RuntimeError as e:
+                if launcher.dynamic_shapes_policy == 'dynamic':
+                    raise e
+                self.is_dynamic = False
+                self.exec_network = None
+        if not self.dynamic_inputs:
+            self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
         self.update_input_output_info(model_prefix)
         self.input_feeder = InputFeeder(
             self.model_info.get('inputs', []), self.inputs, self.input_shape, self.fit_to_input)
 
+    def reshape_net(self, shape):
+        if self.is_dynamic:
+            return
+        if hasattr(self, 'exec_network') and self.exec_network is not None:
+            del self.exec_network
+        self.network.reshape(shape)
+        self.dynamic_inputs, self.partial_shapes = self.launcher.get_dynamic_inputs(self.network)
+        if not self.is_dynamic and self.dynamic_inputs:
+            return
+        self.exec_network = self.launcher.load_network(self.network, self.launcher.device)
+
     def load_model(self, network_info, launcher, model_prefix=None, log=False):
         self.network = launcher.read_network(str(network_info['model']), str(network_info['weights']))
-        self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
-        self.launcher = launcher
-        self.update_input_output_info(model_prefix)
-        self.input_feeder = InputFeeder(
-            self.model_info.get('inputs', []), self.inputs, self.input_shape, self.fit_to_input)
+        self.load_network(self.network, launcher, model_prefix)
         if log:
             self.print_input_output_info()
 
@@ -451,12 +473,14 @@ class DLSDKModelMixin:
         for name, input_info in network_inputs.items():
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(input_info.precision))
-            print_info('\tshape {}\n'.format(input_info.shape))
+            print_info('\tshape {}\n'.format(
+                input_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
         print_info('{} - Output info'.format(self.default_model_name))
         for name, output_info in network_outputs.items():
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(output_info.precision))
-            print_info('\tshape: {}\n'.format(output_info.shape))
+            print_info('\tshape: {}\n'.format(
+                output_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
 
     def update_input_output_info(self, model_prefix):
         def generate_name(prefix, with_prefix, layer_name):
@@ -870,128 +894,3 @@ class MTCNNEvaluator(BaseEvaluator):
             self.select_dataset('')
         details.update(self.dataset.send_annotation_info(self.dataset_config[0]))
         return details
-
-
-def calibrate_predictions(previous_stage_predictions, out, threshold, outputs_mapping, iou_type=None):
-    score = out[0][outputs_mapping['probability_out']][:, 1]
-    pass_t = np.where(score > 0.7)[0]
-    removed_boxes = [i for i in range(previous_stage_predictions[0].size) if i not in pass_t]
-    previous_stage_predictions[0].remove(removed_boxes)
-    previous_stage_predictions[0].scores = score[pass_t]
-    bboxes = np.c_[
-        previous_stage_predictions[0].x_mins, previous_stage_predictions[0].y_mins,
-        previous_stage_predictions[0].x_maxs, previous_stage_predictions[0].y_maxs,
-        previous_stage_predictions[0].scores
-    ]
-    mv = out[0][outputs_mapping['region_out']][pass_t]
-    if iou_type:
-        previous_stage_predictions[0], peek = nms(previous_stage_predictions[0], threshold, iou_type)
-        bboxes = np.c_[
-            previous_stage_predictions[0].x_mins, previous_stage_predictions[0].y_mins,
-            previous_stage_predictions[0].x_maxs, previous_stage_predictions[0].y_maxs,
-            previous_stage_predictions[0].scores
-        ]
-        mv = mv[np.sort(peek).astype(int)]
-    x_mins, y_mins, x_maxs, y_maxs, _ = bbreg(bboxes, mv.T).T
-    previous_stage_predictions[0].x_mins = x_mins
-    previous_stage_predictions[0].y_mins = y_mins
-    previous_stage_predictions[0].x_maxs = x_maxs
-    previous_stage_predictions[0].y_maxs = y_maxs
-    return previous_stage_predictions
-
-def nms(prediction, threshold, iou_type):
-    bboxes = np.c_[prediction.x_mins, prediction.y_mins, prediction.x_maxs, prediction.y_maxs, prediction.scores]
-    peek = MTCNNPAdapter.nms(bboxes, threshold, iou_type)
-    prediction.remove([i for i in range(prediction.size) if i not in peek])
-    return prediction, peek
-
-def bbreg(boundingbox, reg):
-    reg = reg.T
-    # calibrate bounding boxes
-    w = boundingbox[:, 2] - boundingbox[:, 0] + 1
-    h = boundingbox[:, 3] - boundingbox[:, 1] + 1
-    bb0 = boundingbox[:, 0] + reg[:, 0] * w
-    bb1 = boundingbox[:, 1] + reg[:, 1] * h
-    bb2 = boundingbox[:, 2] + reg[:, 2] * w
-    bb3 = boundingbox[:, 3] + reg[:, 3] * h
-    boundingbox[:, 0:4] = np.array([bb0, bb1, bb2, bb3]).T
-    return boundingbox
-
-def filter_valid(dy, edy, dx, edx, y, ey, x, ex, tmpw, tmph):
-    mask = np.ones(len(tmph))
-    tmp_ys_len = (edy + 1) - dy
-    tmp_xs_len = (edx + 1) - dx
-    img_ys_len = (ey + 1) - y
-    img_xs_len = (ex + 1) - x
-    mask = np.logical_and(mask, np.logical_and(tmph > 0, tmpw > 0))
-    mask = np.logical_and(mask, np.logical_and(tmp_ys_len > 0, tmp_xs_len > 0))
-    mask = np.logical_and(mask, np.logical_and(img_xs_len > 0, img_ys_len > 0))
-    mask = np.logical_and(mask, np.logical_and(tmp_xs_len == img_xs_len, tmp_ys_len == img_ys_len))
-    return dy[mask], edy[mask], dx[mask], edx[mask], y[mask], ey[mask], x[mask], ex[mask], tmpw[mask], tmph[mask], mask
-
-def pad(boxesA, h, w):
-    boxes = boxesA.copy()
-    tmph = boxes[:, 3] - boxes[:, 1] + 1
-    tmpw = boxes[:, 2] - boxes[:, 0] + 1
-    numbox = boxes.shape[0]
-    dx = np.ones(numbox)
-    dy = np.ones(numbox)
-    edx = tmpw
-    edy = tmph
-    x = boxes[:, 0:1][:, 0]
-    y = boxes[:, 1:2][:, 0]
-    ex = boxes[:, 2:3][:, 0]
-    ey = boxes[:, 3:4][:, 0]
-    tmp = np.where(ex > w)[0]
-    if tmp.shape[0] != 0:
-        edx[tmp] = -ex[tmp] + w - 1 + tmpw[tmp]
-        ex[tmp] = w - 1
-    tmp = np.where(ey > h)[0]
-    if tmp.shape[0] != 0:
-        edy[tmp] = -ey[tmp] + h - 1 + tmph[tmp]
-        ey[tmp] = h - 1
-    tmp = np.where(x < 1)[0]
-    if tmp.shape[0] != 0:
-        dx[tmp] = 2 - x[tmp]
-        x[tmp] = np.ones_like(x[tmp])
-    tmp = np.where(y < 1)[0]
-    if tmp.shape[0] != 0:
-        dy[tmp] = 2 - y[tmp]
-        y[tmp] = np.ones_like(y[tmp])
-    # for python index from 0, while matlab from 1
-    dy, dx = np.maximum(0, dy - 1), np.maximum(0, dx - 1)
-    y = np.maximum(0, y - 1)
-    x = np.maximum(0, x - 1)
-    edy = np.maximum(0, edy - 1)
-    edx = np.maximum(0, edx - 1)
-    ey = np.maximum(0, ey - 1)
-    ex = np.maximum(0, ex - 1)
-    return filter_valid(dy, edy, dx, edx, y, ey, x, ex, tmpw, tmph)
-
-def rerec(bboxA):
-    w = bboxA[:, 2] - bboxA[:, 0]
-    h = bboxA[:, 3] - bboxA[:, 1]
-    max_side = np.maximum(w, h).T
-    bboxA[:, 0] = bboxA[:, 0] + w * 0.5 - max_side * 0.5
-    bboxA[:, 1] = bboxA[:, 1] + h * 0.5 - max_side * 0.5
-    bboxA[:, 2:4] = bboxA[:, 0:2] + np.repeat([max_side], 2, axis=0).T
-    return bboxA
-
-def cut_roi(image, prediction, dst_size, include_bound=True):
-    bboxes = np.c_[prediction.x_mins, prediction.y_mins, prediction.x_maxs, prediction.y_maxs, prediction.scores]
-    img = image.data
-    bboxes = rerec(bboxes)
-    bboxes[:, 0:4] = np.fix(bboxes[:, 0:4])
-    dy, edy, dx, edx, y, ey, x, ex, tmpw, tmph, mask = pad(bboxes, *img.shape[:2])
-    bboxes = bboxes[mask]
-    numbox = bboxes.shape[0]
-    tempimg = np.zeros((numbox, dst_size, dst_size, 3))
-    for k in range(numbox):
-        tmp_k_h, tmp_k_w = int(tmph[k]) + int(include_bound), int(tmpw[k]) + int(include_bound)
-        tmp = np.zeros((tmp_k_h, tmp_k_w, 3))
-        tmp_ys, tmp_xs = slice(int(dy[k]), int(edy[k]) + 1), slice(int(dx[k]), int(edx[k]) + 1)
-        img_ys, img_xs = slice(int(y[k]), int(ey[k]) + 1), slice(int(x[k]), int(ex[k]) + 1)
-        tmp[tmp_ys, tmp_xs] = img[img_ys, img_xs]
-        tempimg[k, :, :, :] = cv2.resize(tmp, (dst_size, dst_size))
-    image.data = tempimg
-    return image
