+++ tools/accuracy_checker/accuracy_checker/evaluators/custom_evaluators/sequential_action_recognition_evaluator.py
@@ -319,15 +319,13 @@ class BaseModel:
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(input_info.precision))
             print_info('\tshape: {}\n'.format(
-                input_info.shape if name not in self.partial_shapes else self.partial_shapes[name])
-            )
+                input_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
         print_info('{} - Output info'.format(self.default_model_suffix))
         for name, output_info in network_outputs.items():
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(output_info.precision))
             print_info('\tshape: {}\n'.format(
-                output_info.shape if name not in self.partial_shapes else self.partial_shapes[name]
-            ))
+                output_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
 
 
 def create_encoder(model_config, launcher, delayed_model_loading=False):
@@ -468,7 +466,7 @@ class EncoderDLSDKModel(BaseModel):
         if hasattr(self, 'exec_network') and self.exec_network is not None:
             del self.exec_network
         self.network.reshape(shape)
-        self.dynamic_inputs, self.partial_shapes = self.launcher._get_dynamic_inputs(self.network)
+        self.dynamic_inputs, self.partial_shapes = self.launcher.get_dynamic_inputs(self.network)
         if not self.is_dynamic and self.dynamic_inputs:
             return
         self.exec_network = self.launcher.load_network(self.network, self.launcher.device)
@@ -533,7 +531,7 @@ class EncoderDLSDKModel(BaseModel):
 
     def load_network(self, network, launcher):
         self.network = network
-        self.dynamic_inputs, self.partial_shapes = launcher._get_dynamic_inputs(self.network)
+        self.dynamic_inputs, self.partial_shapes = launcher.get_dynamic_inputs(self.network)
         if self.dynamic_inputs and launcher.dynamic_shapes_policy in ['dynamic', 'default']:
             try:
                 self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
@@ -653,7 +651,7 @@ class DecoderDLSDKModel(BaseModel):
 
     def load_network(self, network, launcher):
         self.network = network
-        self.dynamic_inputs, self.partial_shapes = launcher._get_dynamic_inputs(self.network)
+        self.dynamic_inputs, self.partial_shapes = launcher.get_dynamic_inputs(self.network)
         if self.dynamic_inputs and launcher.dynamic_shapes_policy in ['dynamic', 'default']:
             try:
                 self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
@@ -670,7 +668,7 @@ class DecoderDLSDKModel(BaseModel):
         if hasattr(self, 'exec_network') and self.exec_network is not None:
             del self.exec_network
         self.network.reshape(shape)
-        self.dynamic_inputs, self.partial_shapes = self.launcher._get_dynamic_inputs(self.network)
+        self.dynamic_inputs, self.partial_shapes = self.launcher.get_dynamic_inputs(self.network)
         if not self.is_dynamic and self.dynamic_inputs:
             return
         self.exec_network = self.launcher.load_network(self.network, self.launcher.device)
