+++ tools/accuracy_checker/accuracy_checker/evaluators/custom_evaluators/sr_evaluator.py
@@ -318,15 +318,13 @@ class BaseDLSDKModel:
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(input_info.precision))
             print_info('\tshape: {}\n'.format(
-                input_info.shape if name not in self.partial_shapes else self.partial_shapes[name])
-            )
+                input_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
         print_info('{} - Output info'.format(self.default_model_suffix))
         for name, output_info in network_outputs.items():
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(output_info.precision))
             print_info('\tshape: {}\n'.format(
-                output_info.shape if name not in self.partial_shapes else self.partial_shapes[name]
-            ))
+                output_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
 
     def automatic_model_search(self, network_info):
         model = Path(network_info.get('srmodel', network_info.get('model')))
@@ -365,7 +363,7 @@ class BaseDLSDKModel:
 
     def load_network(self, network, launcher):
         self.network = network
-        self.dynamic_inputs, self.partial_shapes = launcher._get_dynamic_inputs(self.network)
+        self.dynamic_inputs, self.partial_shapes = launcher.get_dynamic_inputs(self.network)
         if self.dynamic_inputs and launcher.dynamic_shapes_policy in ['dynamic', 'default']:
             try:
                 self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
@@ -385,7 +383,7 @@ class BaseDLSDKModel:
         if hasattr(self, 'exec_network') and self.exec_network is not None:
             del self.exec_network
         self.network.reshape(shape)
-        self.dynamic_inputs, self.partial_shapes = self.launcher._get_dynamic_inputs(self.network)
+        self.dynamic_inputs, self.partial_shapes = self.launcher.get_dynamic_inputs(self.network)
         if not self.is_dynamic and self.dynamic_inputs:
             return
         self.exec_network = self.launcher.load_network(self.network, self.launcher.device)
@@ -397,7 +395,7 @@ class BaseDLSDKModel:
         model, weights = self.automatic_model_search(network_info)
         if weights is not None:
             self.network = launcher.read_network(str(model), str(weights))
-            self.exec_network = self.load_network(self.network)
+            self.load_network(self.network, launcher)
         else:
             self.exec_network = launcher.ie_core.import_network(str(model))
         self.update_inputs_outputs_info()
