+++ tools/accuracy_checker/accuracy_checker/evaluators/custom_evaluators/text_spotting_evaluator.py
@@ -289,6 +289,7 @@ class BaseModel:
     def __init__(self, network_info, launcher, default_model_suffix, delayed_model_loading=False):
         self.default_model_suffix = default_model_suffix
         self.network_info = network_info
+        self.launcher = launcher
 
     def predict(self, identifiers, input_data):
         raise NotImplementedError
@@ -353,16 +354,40 @@ class BaseModel:
         for name, input_info in network_inputs.items():
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(input_info.precision))
-            print_info('\tshape {}\n'.format(input_info.shape))
+            print_info('\tshape: {}\n'.format(
+                input_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
         print_info('{} - Output info'.format(self.default_model_suffix))
         for name, output_info in network_outputs.items():
             print_info('\tLayer name: {}'.format(name))
             print_info('\tprecision: {}'.format(output_info.precision))
-            print_info('\tshape: {}\n'.format(output_info.shape))
+            print_info('\tshape: {}\n'.format(
+                output_info.shape if name not in self.partial_shapes else self.partial_shapes[name]))
 
     def load_network(self, network, launcher):
         self.network = network
-        self.exec_network = launcher.ie_core.load_network(network, launcher.device)
+        self.dynamic_inputs, self.partial_shapes = launcher.get_dynamic_inputs(self.network)
+        if self.dynamic_inputs and launcher.dynamic_shapes_policy in ['dynamic', 'default']:
+            try:
+                self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
+                self.is_dynamic = True
+            except RuntimeError as e:
+                if launcher.dynamic_shapes_policy == 'dynamic':
+                    raise e
+                self.is_dynamic = False
+                self.exec_network = None
+        if not self.dynamic_inputs:
+            self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
+
+    def reshape_net(self, shape):
+        if self.is_dynamic:
+            return
+        if hasattr(self, 'exec_network') and self.exec_network is not None:
+            del self.exec_network
+        self.network.reshape(shape)
+        self.dynamic_inputs, self.partial_shapes = self.launcher.get_dynamic_inputs(self.network)
+        if not self.is_dynamic and self.dynamic_inputs:
+            return
+        self.exec_network = self.launcher.load_network(self.network, self.launcher.device)
 
     def get_network(self):
         return self.network
@@ -574,6 +599,8 @@ class DetectorDLSDKModel(BaseModel):
                           self.im_info_name: np.array([[input_data.shape[1], input_data.shape[2], 1.0]])}
         else:
             input_data = {self.im_data_name: self.fit_to_input(input_data)}
+        if not self.is_dynamic and self.dynamic_inputs:
+            self.reshape_net({k: v.shape for k, v in input_data.items()})
 
         output = self.exec_network.infer(input_data)
 
@@ -598,7 +625,7 @@ class DetectorDLSDKModel(BaseModel):
         model, weights = self.automatic_model_search(network_info)
         if weights is not None:
             self.network = launcher.read_network(str(model), str(weights))
-            self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
+            self.exec_network = self.load_network(self.network, launcher)
         else:
             self.exec_network = launcher.ie_core.import_network(str(model))
         has_info = hasattr(self.exec_network, 'input_info')
@@ -626,6 +653,8 @@ class RecognizerDLSDKModel(BaseModel):
             self.load_model(network_info, launcher, log=True)
 
     def predict(self, identifiers, input_data):
+        if not self.is_dynamic and self.dynamic_inputs:
+            self.reshape_net({k: v.shape for k, v in input_data.items()})
         return self.exec_network.infer(input_data)
 
     def release(self):
@@ -636,7 +665,7 @@ class RecognizerDLSDKModel(BaseModel):
         model, weights = self.automatic_model_search(network_info)
         if weights is not None:
             self.network = launcher.read_network(str(model), str(weights))
-            self.exec_network = launcher.ie_core.load_network(self.network, launcher.device)
+            self.exec_network = self.load_network(self.network, launcher.device)
         else:
             self.exec_network = launcher.ie_core.import_network(str(model))
         if log:
